# api_docs.yaml
openapi: 3.0.3
info:
  title: Adafruit Pump Control API
  version: 1.0.0
  description: APIs to control a green house IOT devices.
servers:
  - url: http://127.0.0.1:5000
    description: Local development server
...
paths:
  /api/adafruit/pump/on:
    post:
      tags: 
        - Manual mode
      summary: Turn on the pump for a specified duration
      operationId: turnPumpOn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - duration
              properties:
                duration:
                  type: integer
                  minimum: 1
                  description: Duration in seconds to turn on the pump
      responses:
        '200':
          description: Pump successfully turned on
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Pump was turn on for 5 second(s)"
        '400':
          description: Bad request due to missing or invalid duration parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Missing duration parameter"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
  /api/adafruit/servo:
    post:
      tags: 
        - Manual mode
      summary: Open the cover to a specified angle using the servo
      operationId: setServoAngle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - angle
              properties:
                angle:
                  type: integer
                  minimum: 0
                  maximum: 180
                  description: Angle in degrees to open the cover (0–180)
      responses:
        '200':
          description: Servo angle successfully set
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  angle:
                    type: integer
                    example: 90
        '400':
          description: Bad request due to missing or invalid angle parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Missing angle parameter, or angle is not int, or angle is not between 0 and 180"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Internal server error"
  /api/adafruit/fan:
    post:
      tags: 
        - Manual mode
      summary: Turn on the fan with a specified speed
      operationId: setFanSpeed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - speed
              properties:
                speed:
                  type: integer
                  minimum: 0
                  maximum: 255
                  description: Fan speed value (0–255)
      responses:
        '200':
          description: Fan speed successfully set
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  current_fan_speed:
                    type: integer
                    example: 128
        '400':
          description: Bad request due to missing or invalid speed parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Missing speed parameter, speed is not number, speed is not between 0 and 255"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Internal server error"
...